requirements.txt
absl-py==2.2.1
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.11.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.23.0
comm==0.2.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.3.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.0.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.60.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
graphtools==2.1.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.11.1
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.0
jupyter_core==5.6.0
keras==3.9.1
keybert==0.9.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llvmlite==0.45.1
marisa-trie==1.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.6
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.13
namex==0.0.8
narwhals==2.8.0
nest-asyncio==1.5.8
networkx==3.4.2
nltk==3.9.1
numba==0.62.1
numpy==2.1.3
openai==2.6.1
opt_einsum==3.4.0
optree==0.14.1
orjson==3.11.8
packaging==23.2
pandas==2.2.3
parso==0.8.3
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
phate==2.0.0
pillow==11.0.0
pinecone==8.1.0
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.1.0
plotly==6.3.1
preshed==3.0.10
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==5.29.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.17.2
PyGSP==0.6.1
pynndescent==0.5.13
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.2.0
python-dateutil==2.8.2
python-dotenv==1.2.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.27.1
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.1
sentencepiece==0.2.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart_open==7.4.2
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tasklogger==1.2.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.0
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.12
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
Wikipedia-API==0.8.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
